LayoutTests/inspector/elements/set-outer-html-for-xhtml-expected.txt

Tests DOMAgent.setOuterHTML protocol method against an XHTML document.


Running: testSetUp

Running: testChangeCharacterData
Replacing 'Getting involved' with 'Getting not involved'

Wrapper identity: identity
Event AttrModified: div
Event AttrModified: div
Event AttrModified: div
Event AttrRemoved: div
Event AttrRemoved: div
Event CharacterDataModified: 
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting not involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event CharacterDataModified: 
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testChangeAttributes
Replacing '<a href' with '<a foo="bar" href'

Wrapper identity: identity
Event AttrModified: a
Event AttrModified: a
Event AttrRemoved: a
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a foo="bar" href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event AttrModified: a
Event AttrRemoved: a
Event AttrRemoved: a
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testRemoveLastChild
Replacing 'Getting involved' with ''

Wrapper identity: identity
Event NodeRemoved: 
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml"></h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: 
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testSplitNode
Replacing 'Getting involved' with 'Getting</h2><h2>involved'

Wrapper identity: identity
Event NodeInserted: h2
Event NodeInserted: h2
Event NodeRemoved: h2
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting</h2><h2 xmlns="http://www.w3.org/1999/xhtml">involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: h2
Event NodeRemoved: h2
Event NodeRemoved: h2
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testChangeNodeName
Replacing '<h2>Getting involved</h2>' with '<h3>Getting involved</h3>'

Wrapper identity: identity
Event NodeInserted: h3
Event NodeRemoved: h2
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h3 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h3>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: h2
Event NodeRemoved: h3
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p xmlns="http://www.w3.org/1999/xhtml">WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2 xmlns="http://www.w3.org/1999/xhtml">Getting involved</h2>
<p xmlns="http://www.w3.org/1999/xhtml" id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========
